Add initStoryblok func
