+++ apps/store/src/pages/_app.tsx
@@ -3,12 +3,11 @@ import { storyblokInit, apiPlugin } from '@storyblok/react'
 import type { AppPropsWithLayout } from 'next/app'
 import Head from 'next/head'
 import { ThemeProvider } from 'ui'
-import { HeadingBlock } from '@/blocks/HeadingBlock'
-import { PageBlock } from '@/blocks/PageBlock'
 import { GlobalStyles } from '@/lib/GlobalStyles'
 import { useApollo } from '@/services/apollo/client'
 import * as Datadog from '@/services/datadog'
 import { CartContext, useCartContextStore } from '@/services/mockCartService'
+import { initStoryblok } from '@/services/storyblok'
 
 // Enable API mocking
 if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
@@ -17,16 +16,7 @@ if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
 
 Datadog.initRum()
 
-const components = {
-  heading: HeadingBlock,
-  page: PageBlock,
-}
-
-storyblokInit({
-  accessToken: process.env.NEXT_PUBLIC_STORYBLOK_ACCESS_TOKEN,
-  use: [apiPlugin],
-  components,
-})
+initStoryblok()
 
 const MyApp = ({ Component, pageProps }: AppPropsWithLayout) => {
   const apolloClient = useApollo(pageProps)
