--- apps/store/src/services/storyblok/storyblok.ts
@@ -1,5 +1,0 @@
-import { SbBlokData } from '@storyblok/react'
-
-export type SbBaseBlockProps<T> = {
-  blok: SbBlokData & T
-}
